keep cellRenderer clean
